fix(siteread): reject recombined identifiers, and clear a stale entity
Three more from review, all real.

groundedDetail matched the claimed tokens as an unordered SET, so a value
assembled from tokens the block happens to contain separately passed: a
notice printing "24114 Kiel" and "HRB 123456" vouched for the invented
"HRB 24114". The tokens must now appear as a contiguous run, which still
tolerates the punctuation a model rearranges — punctuation is not a token
— while refusing a fabricated legal identifier.

Reading a SECOND website kept the first one's legal identity. Picking an
entity marks those three fields as yours, and prefill deliberately leaves
your own input alone, so one company's registered name survived into
another company's setup. A new read clears them; its census fills them
again.

And --fresh's refusal printed the DSN it rejected, password and all —
stripping the query string leaves user:pass@ intact. It names no
connection at all now, which is the only safe amount of it to say.

sitepagefacts.go crossed the 500-line cap, so the legal-entity lane moved
to sitepageentities.go: a legal page answers that lane and no other page
kind does, and its rules are about attribution rather than evidence.
